feat: Delete data dump from S3 after data load has been successful (#600)
