+++ dbt_platform_helper/providers/load_balancers.py
@@ -1,0 +1,51 @@
+import boto3
+
+
+def find_load_balancer(session: boto3.Session, app: str, env: str) -> str:
+    lb_client = session.client("elbv2")
+
+    describe_response = lb_client.describe_load_balancers()
+    load_balancers = [lb["LoadBalancerArn"] for lb in describe_response["LoadBalancers"]]
+
+    load_balancers = lb_client.describe_tags(ResourceArns=load_balancers)["TagDescriptions"]
+
+    load_balancer_arn = None
+    for lb in load_balancers:
+        tags = {t["Key"]: t["Value"] for t in lb["Tags"]}
+        if tags.get("copilot-application") == app and tags.get("copilot-environment") == env:
+            load_balancer_arn = lb["ResourceArn"]
+
+    if not load_balancer_arn:
+        raise LoadBalancerNotFoundError()
+
+    return load_balancer_arn
+
+
+def find_https_listener(session: boto3.Session, app: str, env: str) -> str:
+    load_balancer_arn = find_load_balancer(session, app, env)
+    lb_client = session.client("elbv2")
+    listeners = lb_client.describe_listeners(LoadBalancerArn=load_balancer_arn)["Listeners"]
+
+    listener_arn = None
+
+    try:
+        listener_arn = next(l["ListenerArn"] for l in listeners if l["Protocol"] == "HTTPS")
+    except StopIteration:
+        pass
+
+    if not listener_arn:
+        raise ListenerNotFoundError()
+
+    return listener_arn
+
+
+class LoadBalancerNotFoundError(Exception):
+    pass
+
+
+class ListenerNotFoundError(Exception):
+    pass
+
+
+class ListenerRuleNotFoundError(Exception):
+    pass
