--- images/tools/database-copy/shell-profile.sh
@@ -1,3 +1,0 @@
-echo "Starting database copy."
-
-pg_dump $SOURCE_DB_CONNECTION | psql $TARGET_DB_CONNECTION main
