--- images/tools/database-copy2/Dockerfile
@@ -1,11 +1,0 @@
-FROM postgres:16
-
-RUN apt update && apt upgrade
-RUN apt install -y curl zip jq
-RUN curl "https://awscli.amazonaws.com/awscli-exe-linux-aarch64.zip" -o "awscliv2.zip"
-RUN unzip awscliv2.zip
-RUN ./aws/install
-
-COPY entrypoint.sh /entrypoint.sh
-
-ENTRYPOINT ["bash", "/entrypoint.sh"]
