+++ tests/platform_helper/test_command_database.py
@@ -2,6 +2,7 @@ from unittest.mock import patch
 
 from click.testing import CliRunner
 
+from dbt_platform_helper.commands.database import copy
 from dbt_platform_helper.commands.database import dump
 from dbt_platform_helper.commands.database import load
 
@@ -14,24 +15,20 @@ def test_command_dump_success(mock_database_copy_object):
     result = runner.invoke(
         dump,
         [
-            "--account-id",
-            "12345",
             "--app",
             "my_app",
-            "--env",
+            "--from",
             "my_env",
             "--database",
             "my_postgres",
-            "--vpc-name",
+            "--from-vpc",
             "my_vpc",
         ],
     )
 
     assert result.exit_code == 0
-    mock_database_copy_object.assert_called_once_with(
-        "12345", "my_app", "my_env", "my_postgres", "my_vpc"
-    )
-    mock_database_copy_instance.dump.assert_called_once_with()
+    mock_database_copy_object.assert_called_once_with("my_app", "my_postgres")
+    mock_database_copy_instance.dump.assert_called_once_with("my_env", "my_vpc")
 
 
 @patch("dbt_platform_helper.commands.database.DatabaseCopy")
@@ -41,21 +38,199 @@ def test_command_load_success(mock_database_copy_object):
     result = runner.invoke(
         load,
         [
-            "--account-id",
-            "12345",
             "--app",
             "my_app",
-            "--env",
+            "--to",
             "my_env",
             "--database",
             "my_postgres",
-            "--vpc-name",
+            "--to-vpc",
             "my_vpc",
         ],
     )
 
     assert result.exit_code == 0
-    mock_database_copy_object.assert_called_once_with(
-        "12345", "my_app", "my_env", "my_postgres", "my_vpc"
+    mock_database_copy_object.assert_called_once_with("my_app", "my_postgres", False)
+    mock_database_copy_instance.load.assert_called_once_with("my_env", "my_vpc")
+
+
+@patch("dbt_platform_helper.commands.database.DatabaseCopy")
+def test_command_load_success_with_auto_approve(mock_database_copy_object):
+    mock_database_copy_instance = mock_database_copy_object.return_value
+    runner = CliRunner()
+    result = runner.invoke(
+        load,
+        [
+            "--app",
+            "my_app",
+            "--to",
+            "my_env",
+            "--database",
+            "my_postgres",
+            "--to-vpc",
+            "my_vpc",
+            "--auto-approve",
+        ],
     )
-    mock_database_copy_instance.load.assert_called_once_with()
+
+    assert result.exit_code == 0
+    mock_database_copy_object.assert_called_once_with("my_app", "my_postgres", True)
+    mock_database_copy_instance.load.assert_called_once_with("my_env", "my_vpc")
+
+
+@patch("dbt_platform_helper.commands.database.DatabaseCopy")
+def test_command_copy_success(mock_database_copy_object):
+    mock_database_copy_instance = mock_database_copy_object.return_value
+    runner = CliRunner()
+    result = runner.invoke(
+        copy,
+        [
+            "--app",
+            "my_app",
+            "--from",
+            "my_prod_env",
+            "--to",
+            "my_hotfix_env",
+            "--database",
+            "my_postgres",
+            "--from-vpc",
+            "my_from_vpc",
+            "--to-vpc",
+            "my_to_vpc",
+        ],
+    )
+
+    assert result.exit_code == 0
+    mock_database_copy_object.assert_called_once_with("my_app", "my_postgres", False)
+    mock_database_copy_instance.copy.assert_called_once_with(
+        "my_prod_env",
+        "my_hotfix_env",
+        "my_from_vpc",
+        "my_to_vpc",
+        ("web",),
+        "default",
+        False,
+    )
+
+
+@patch("dbt_platform_helper.commands.database.DatabaseCopy")
+def test_command_copy_success_with_auto_approve(mock_database_copy_object):
+    mock_database_copy_instance = mock_database_copy_object.return_value
+    runner = CliRunner()
+    result = runner.invoke(
+        copy,
+        [
+            "--app",
+            "my_app",
+            "--from",
+            "my_prod_env",
+            "--to",
+            "my_hotfix_env",
+            "--database",
+            "my_postgres",
+            "--from-vpc",
+            "my_from_vpc",
+            "--to-vpc",
+            "my_to_vpc",
+            "--auto-approve",
+            "--svc",
+            "other",
+            "--svc",
+            "service",
+            "--template",
+            "migration",
+        ],
+    )
+
+    assert result.exit_code == 0
+    mock_database_copy_object.assert_called_once_with("my_app", "my_postgres", True)
+    mock_database_copy_instance.copy.assert_called_once_with(
+        "my_prod_env",
+        "my_hotfix_env",
+        "my_from_vpc",
+        "my_to_vpc",
+        ("other", "service"),
+        "migration",
+        False,
+    )
+
+
+@patch("dbt_platform_helper.commands.database.DatabaseCopy")
+def test_command_copy_success_with_no_maintenance_page(mock_database_copy_object):
+    mock_database_copy_instance = mock_database_copy_object.return_value
+    runner = CliRunner()
+    result = runner.invoke(
+        copy,
+        [
+            "--app",
+            "my_app",
+            "--from",
+            "my_prod_env",
+            "--to",
+            "my_hotfix_env",
+            "--database",
+            "my_postgres",
+            "--from-vpc",
+            "my_from_vpc",
+            "--to-vpc",
+            "my_to_vpc",
+            "--auto-approve",
+            "--svc",
+            "other",
+            "--svc",
+            "service",
+            "--no-maintenance-page",
+        ],
+    )
+
+    assert result.exit_code == 0
+    mock_database_copy_object.assert_called_once_with("my_app", "my_postgres", True)
+    mock_database_copy_instance.copy.assert_called_once_with(
+        "my_prod_env",
+        "my_hotfix_env",
+        "my_from_vpc",
+        "my_to_vpc",
+        ("other", "service"),
+        "default",
+        True,
+    )
+
+
+@patch("dbt_platform_helper.commands.database.DatabaseCopy")
+def test_command_copy_success_with_maintenance_page(mock_database_copy_object):
+    mock_database_copy_instance = mock_database_copy_object.return_value
+    runner = CliRunner()
+    result = runner.invoke(
+        copy,
+        [
+            "--app",
+            "my_app",
+            "--from",
+            "my_prod_env",
+            "--to",
+            "my_hotfix_env",
+            "--database",
+            "my_postgres",
+            "--from-vpc",
+            "my_from_vpc",
+            "--to-vpc",
+            "my_to_vpc",
+            "--auto-approve",
+            "--svc",
+            "other",
+            "--svc",
+            "service",
+        ],
+    )
+
+    assert result.exit_code == 0
+    mock_database_copy_object.assert_called_once_with("my_app", "my_postgres", True)
+    mock_database_copy_instance.copy.assert_called_once_with(
+        "my_prod_env",
+        "my_hotfix_env",
+        "my_from_vpc",
+        "my_to_vpc",
+        ("other", "service"),
+        "default",
+        False,
+    )
